Move task list logic outside of WpcomChecklist render tree.
Task list logic was stuck inside the WpcomChecklist render tree,
forcing checklist instances to look at the list of children to determine
completion status. This also made it difficult to determine whether
the checklist should be rendered without actually rendering it.

This commit refactors the task list into a class that's instanced
by WpcomChecklist and attached as a prop to the checklist component
in its render tree.
This instance serves as the single source of truth on the list of
tasks, the completion status of each task, and the overall completion
status for the task list.